split scrape_mt into common and mt specific
